+++ opticksgl/go.sh
@@ -6,7 +6,12 @@ sdir=$(pwd)
 name=$(basename $sdir)
 
 bdir=$(opticks-prefix)/build/$name
-rm -rf $bdir 
+
+if [ "$1" == "clean" ]; then
+   echo $0 $1 : remove bdir $bdir
+   rm -rf $bdir 
+fi
+
 mkdir -p $bdir && cd $bdir && pwd 
 
 cmake $sdir \
@@ -19,4 +24,9 @@ cmake $sdir \
 
 make
 make install   
+
+opticks-t $bdir
+
+
+
